refactor(acp): modularize monolithic acpClient into specialized files
Decomposed the massive `acpClient.ts` and its test file into smaller, focused modules to improve maintainability and isolate concerns. This completes Phase 1 of the ACP refactoring plan. Additionally, all files in the `acp` folder have been renamed to share the `acp` prefix for better consistency.

- **Source Code Restructuring**:
  - Extracted `runAcpClient` to `acpStdioTransport.ts` for handling raw I/O.
  - Extracted `GeminiAgent` to `acpRpcDispatcher.ts` to act as pure RPC handler.
  - Created `acpSessionManager.ts` to manage concurrent user sessions.
  - Extracted `Session` class to `acpSession.ts` to manage individual chat turns.
  - Moved shared helpers and schemas to `acpUtils.ts`.
- **File Renaming**: Renamed all files in the `acp` directory to share the `acp` prefix (e.g., `commandHandler.ts` -> `acpCommandHandler.ts`, `fileSystemService.ts` -> `acpFileSystemService.ts`) to avoid name collisions and improve organization.
- **Test Split**: Decomposed the massive `acpClient.test.ts` into corresponding test files: `acpRpcDispatcher.test.ts`, `acpSessionManager.test.ts`, and `acpSession.test.ts`.
- **Integrations**: Updated imports in `gemini.tsx` and `acpResume.test.ts` to use the new file paths.
- **Verification**: Verified that all tests pass and the full build (`npm run build`) succeeds after resolving initial linting and compilation errors.
